Module google_orgpolicy2::api

Structs

FolderConstraintListCall
Lists Constraints that could be applied on the specified resource.

FolderMethods
A builder providing access to all methods supported on folder resources. It is not used directly, but through the OrgPolicyAPI hub.

FolderPolicyCreateCall
Creates a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint does not exist. Returns a google.rpc.Status with google.rpc.Code.ALREADY_EXISTS if the policy already exists on the given Cloud resource.

FolderPolicyDeleteCall
Deletes a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or Org Policy does not exist.

FolderPolicyGetCall
Gets a Policy on a resource. If no Policy is set on the resource, NOT_FOUND is returned. The etag value can be used with UpdatePolicy() to update a Policy during read-modify-write.

FolderPolicyGetEffectivePolicyCall
Gets the effective Policy on a resource. This is the result of merging Policies in the resource hierarchy and evaluating conditions. The returned Policy will not have an etag or condition set because it is a computed Policy across multiple resources. Subtrees of Resource Manager resource hierarchy with 'under:' prefix will not be expanded.

FolderPolicyListCall
Retrieves all of the Policies that exist on a particular resource.

FolderPolicyPatchCall
Updates a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or the policy do not exist. Returns a google.rpc.Status with google.rpc.Code.ABORTED if the etag supplied in the request does not match the persisted etag of the policy Note: the supplied policy will perform a full overwrite of all fields.

GoogleCloudOrgpolicyV2Constraint
A constraint describes a way to restrict resource's configuration. For example, you could enforce a constraint that controls which cloud services can be activated across an organization, or whether a Compute Engine instance can have serial port connections established. Constraints can be configured by the organization's policy adminstrator to fit the needs of the organzation by setting a policy that includes constraints at different locations in the organization's resource hierarchy. Policies are inherited down the resource hierarchy from higher levels, but can also be overridden. For details about the inheritance rules please read about policies. Constraints have a default behavior determined by the constraint_default field, which is the enforcement behavior that is used in the absence of a policy being defined or inherited for the resource in question.

GoogleCloudOrgpolicyV2ConstraintBooleanConstraint
A Constraint that is either enforced or not. For example a constraint constraints/compute.disableSerialPortAccess. If it is enforced on a VM instance, serial port connections will not be opened to that instance.

GoogleCloudOrgpolicyV2ConstraintListConstraint
A Constraint that allows or disallows a list of string values, which are configured by an Organization's policy administrator with a Policy.

GoogleCloudOrgpolicyV2ListConstraintsResponse
The response returned from the ListConstraints method.

GoogleCloudOrgpolicyV2ListPoliciesResponse
The response returned from the ListPolicies method. It will be empty if no Policies are set on the resource.

GoogleCloudOrgpolicyV2Policy
Defines a Cloud Organization Policy which is used to specify Constraints for configurations of Cloud Platform resources.

GoogleCloudOrgpolicyV2PolicySpec
Defines a Cloud Organization PolicySpec which is used to specify Constraints for configurations of Cloud Platform resources.

GoogleCloudOrgpolicyV2PolicySpecPolicyRule
A rule used to express this policy.

GoogleCloudOrgpolicyV2PolicySpecPolicyRuleStringValues
A message that holds specific allowed and denied values. This message can define specific values and subtrees of Cloud Resource Manager resource hierarchy (Organizations, Folders, Projects) that are allowed or denied. This is achieved by using the under: and optional is: prefixes. The under: prefix is used to denote resource subtree values. The is: prefix is used to denote specific values, and is required only if the value contains a ":". Values prefixed with "is:" are treated the same as values with no prefix. Ancestry subtrees must be in one of the following formats: - "projects/", e.g. "projects/tokyo-rain-123" - "folders/", e.g. "folders/1234" - "organizations/", e.g. "organizations/1234" The supports_under field of the associated Constraint defines whether ancestry prefixes can be used.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleTypeExpr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

OrgPolicyAPI
Central instance to access all OrgPolicyAPI related resource activities

OrganizationConstraintListCall
Lists Constraints that could be applied on the specified resource.

OrganizationMethods
A builder providing access to all methods supported on organization resources. It is not used directly, but through the OrgPolicyAPI hub.

OrganizationPolicyCreateCall
Creates a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint does not exist. Returns a google.rpc.Status with google.rpc.Code.ALREADY_EXISTS if the policy already exists on the given Cloud resource.

OrganizationPolicyDeleteCall
Deletes a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or Org Policy does not exist.

OrganizationPolicyGetCall
Gets a Policy on a resource. If no Policy is set on the resource, NOT_FOUND is returned. The etag value can be used with UpdatePolicy() to update a Policy during read-modify-write.

OrganizationPolicyGetEffectivePolicyCall
Gets the effective Policy on a resource. This is the result of merging Policies in the resource hierarchy and evaluating conditions. The returned Policy will not have an etag or condition set because it is a computed Policy across multiple resources. Subtrees of Resource Manager resource hierarchy with 'under:' prefix will not be expanded.

OrganizationPolicyListCall
Retrieves all of the Policies that exist on a particular resource.

OrganizationPolicyPatchCall
Updates a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or the policy do not exist. Returns a google.rpc.Status with google.rpc.Code.ABORTED if the etag supplied in the request does not match the persisted etag of the policy Note: the supplied policy will perform a full overwrite of all fields.

ProjectConstraintListCall
Lists Constraints that could be applied on the specified resource.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the OrgPolicyAPI hub.

ProjectPolicyCreateCall
Creates a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint does not exist. Returns a google.rpc.Status with google.rpc.Code.ALREADY_EXISTS if the policy already exists on the given Cloud resource.

ProjectPolicyDeleteCall
Deletes a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or Org Policy does not exist.

ProjectPolicyGetCall
Gets a Policy on a resource. If no Policy is set on the resource, NOT_FOUND is returned. The etag value can be used with UpdatePolicy() to update a Policy during read-modify-write.

ProjectPolicyGetEffectivePolicyCall
Gets the effective Policy on a resource. This is the result of merging Policies in the resource hierarchy and evaluating conditions. The returned Policy will not have an etag or condition set because it is a computed Policy across multiple resources. Subtrees of Resource Manager resource hierarchy with 'under:' prefix will not be expanded.

ProjectPolicyListCall
Retrieves all of the Policies that exist on a particular resource.

ProjectPolicyPatchCall
Updates a Policy. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or the policy do not exist. Returns a google.rpc.Status with google.rpc.Code.ABORTED if the etag supplied in the request does not match the persisted etag of the policy Note: the supplied policy will perform a full overwrite of all fields.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.